azure_functions/OpenAiProcess/requirements.txt
# Azure Functions Core
azure-functions>=1.18.0
azure-functions-worker>=1.0.0

# Azure Services
azure-ai-documentintelligence>=1.0.0
azure-core>=1.29.0
azure-identity>=1.15.0
azure-servicebus>=7.11.0
azure-storage-blob>=12.19.0

# OpenAI
openai>=1.12.0

# Data Processing
tiktoken>=0.5.2

# Environment and Configuration
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
typing-extensions>=4.8.0

# Logging and Monitoring
opencensus-ext-azure>=1.1.13
opencensus-ext-logging>=0.1.1

# JSON and Data Handling
jsonschema>=4.20.0

# Date and Time
python-dateutil>=2.8.2

# Path handling
pathlib2>=2.3.7; python_version < '3.4'

# UUID generation
uuid>=1.30; python_version < '3.7'